+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 
 ## [Unreleased](https://github.com/ixocreate/type-package/compare/0.1.0...develop)
+### Fixed
+- Quill linebreak calculation 
 
 ## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/type-package/compare/master...0.1.0)